+++ include/GMLIB/Files/Language/JsonLanguage.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "GMLIB/Files/JsonConfig.h"
 #include "GMLIB/GMLIB.h"
 
 namespace GMLIB::Files {
