+++ include/GMLIB/include_lib.h
@@ -11,12 +11,12 @@
 #include "GMLIB/Event/Level/WeatherUpdateEvent.h"
 #include "GMLIB/Event/Player/PlayerLoginEvent.h"
 #include "GMLIB/Event/Player/PlayerSleepEvent.h"
-#include "GMLIB/Files/JsonConfig.h"
 #include "GMLIB/Files/JsonFile.h"
-#include "GMLIB/Files/JsonLanguage.h"
-#include "GMLIB/Files/LangLanguage.h"
-#include "GMLIB/Files/McLang.h"
-#include "GMLIB/Files/ResourceLanguage.h"
+#include "GMLIB/Files/Config/JsonConfig.h"
+#include "GMLIB/Files/Language/JsonLanguage.h"
+#include "GMLIB/Files/Language/LangLanguage.h"
+#include "GMLIB/Files/Language/McLang.h"
+#include "GMLIB/Files/Language/ResourceLanguage.h"
 #include "GMLIB/Mod/CustomBiome.h"
 #include "GMLIB/Mod/CustomDamageCause.h"
 #include "GMLIB/Mod/CustomPacks.h"
